- Made the function, which calculates velocities and accelerations for   each motor, as a public static function. So that I can test/use it   somewhere else. Its purpose: take relative movements of each axis in   steps and also values of velocity and acceleration of the axis, which   will move the furthesе. output: values of velocities and accelerations   for each axis OR the error (either input is invalid or some math   broke) Agreements: 1) If the velocity is non-zero, but less than the minimal motor    velocity, we quantize it to the minimal motor velocity 2) If the acceleration is non-zero, but less than the minimal motor    acceleration, we quantize it to the minimal motor acceleration 3) If the angle is lower than the minimal possible angle, it quantizes    down to 0 angles (no movement)
- Changed outData type to queue. I thought it could help with the
  problem of queue stuffing, but does not seem, like it helps
- Tests for computePrepareMove. Cannot run all of the tests at once,
  because of the problem with addDataToOutQueue
- computePrepareMove's inputs should be in the following format:
  movements are in steps; velocity and accleration are in the range [0,
  1] -- percent of the maximum allowed value of the corresponding
     parameters

+++ Params.h
@@ -20,8 +20,8 @@ struct MoveParams
     ParamsStatus status = ParamsStatus::OK;
     String errorMsg;
     double movementUnits[N] = {};
-    uint32_t speed = 0;
-    uint32_t acceleration = 0;
+    double speed = 0; // It should be in percent [0, 1]
+    double acceleration = 0; // It should be in percent [0, 1]
 };
 
 struct MotorIndices
